v0.1: fix: Prettier Configuration + Webview Splitting
