+++ package.json
@@ -13,7 +13,10 @@
   ],
   "main": "./dist/extension.js",
   "activationEvents": [],
-  "extensionKind": ["ui", "workspace"],
+  "extensionKind": [
+    "ui",
+    "workspace"
+  ],
   "capabilities": {
     "virtualWorkspaces": true
   },
@@ -98,7 +101,9 @@
     "watch:tsc": "tsc --noEmit --watch --project tsconfig.json",
     "package": "npm run check-types && npm run lint && node esbuild.js --production",
     "check-types": "tsc --noEmit",
-    "lint": "eslint src"
+    "lint": "eslint src",
+    "format": "prettier --write .",
+    "format:check": "prettier --check ."
   },
   "devDependencies": {
     "@types/node": "20.x",
@@ -107,7 +112,10 @@
     "@typescript-eslint/parser": "^8.7.0",
     "esbuild": "^0.24.0",
     "eslint": "^9.13.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.2.1",
     "npm-run-all": "^4.1.5",
+    "prettier": "^3.3.3",
     "typescript": "^5.6.3"
   },
   "license": "MIT",
