+++ src/constants/index.ts
@@ -4,10 +4,10 @@ export const DEFAULT_ENCODING = 'utf-8';
 
 // Tree drawing symbols used for the GitIngest-style formatter and parser
 export const TREE_SYMBOLS = {
-    VERTICAL: '│',        // vertical guide
-    BRANCH: '├── ',       // branch connector (with trailing space)
-    LAST: '└── ',         // last connector (with trailing space)
-    INDENT: '    '        // 4-space indent used between levels
+    VERTICAL: '│', // vertical guide
+    BRANCH: '├── ', // branch connector (with trailing space)
+    LAST: '└── ', // last connector (with trailing space)
+    INDENT: '    ', // 4-space indent used between levels
 } as const;
 
 export const ERROR_MESSAGES = {
@@ -15,7 +15,7 @@ export const ERROR_MESSAGES = {
     INVALID_DIRECTORY: 'Please select a valid folder.',
     PERMISSION_DENIED: 'Permission denied or file inaccessible.',
     EMPTY_STRUCTURE: 'Plain Text structure cannot be empty.',
-    TARGET_REQUIRED: 'Target directory is required.'
+    TARGET_REQUIRED: 'Target directory is required.',
 } as const;
 
 /*
