+++ src/extension.ts
@@ -2,10 +2,10 @@ import * as vscode from 'vscode';
 import { registerCommands } from './commands';
 
 export function activate(context: vscode.ExtensionContext): void {
-	registerCommands(context);
+    registerCommands(context);
 }
 
-export function deactivate(): void { }
+export function deactivate(): void {}
 
 /*
  * Copyright (c) 2025 Shrey Purohit.
