+++ src/services/formatters/gitignestFormatter.ts
@@ -15,10 +15,16 @@ export class GitignestFormatter extends BaseFormatter {
         return entries
             .map(([name, value], idx) => {
                 const isLast = idx === entries.length - 1;
-                const connector = isRoot ? TREE_SYMBOLS.LAST : (isLast ? TREE_SYMBOLS.LAST : TREE_SYMBOLS.BRANCH);
+                const connector = isRoot
+                    ? TREE_SYMBOLS.LAST
+                    : isLast
+                      ? TREE_SYMBOLS.LAST
+                      : TREE_SYMBOLS.BRANCH;
                 const isDir = typeof value === 'object' && value !== null;
                 const displayName = !isDir
-                    ? ((value as string) === 'file' || (value as string).trim() === '' ? name : `${name}.${value as string}`)
+                    ? (value as string) === 'file' || (value as string).trim() === ''
+                        ? name
+                        : `${name}.${value as string}`
                     : `${name}/`;
                 const line = `${isRoot ? '' : prefix}${connector}${displayName}`;
 
@@ -28,7 +34,10 @@ export class GitignestFormatter extends BaseFormatter {
 
                 const childPrefix = isRoot
                     ? TREE_SYMBOLS.INDENT
-                    : prefix + (isLast ? TREE_SYMBOLS.INDENT : `${TREE_SYMBOLS.VERTICAL}${TREE_SYMBOLS.INDENT.slice(1)}`);
+                    : prefix +
+                      (isLast
+                          ? TREE_SYMBOLS.INDENT
+                          : `${TREE_SYMBOLS.VERTICAL}${TREE_SYMBOLS.INDENT.slice(1)}`);
                 const nested = this.formatStructure(value as FolderStructure, childPrefix);
                 return `${line}\n${nested}`;
             })
