+++ src/utils/parser.ts
@@ -1,5 +1,5 @@
-import { TREE_SYMBOLS } from "../constants";
-import { TreeNode } from "../types";
+import { TREE_SYMBOLS } from '../constants';
+import { TreeNode } from '../types';
 
 class LineParser {
     static countIndentation(rawLine: string): number {
@@ -8,7 +8,10 @@ class LineParser {
         let level = 0;
         while (i < rawLine.length) {
             const next = rawLine.slice(i);
-            if (next.startsWith(TREE_SYMBOLS.VERTICAL + TREE_SYMBOLS.INDENT.slice(1)) || next.startsWith('|' + TREE_SYMBOLS.INDENT.slice(1))) {
+            if (
+                next.startsWith(TREE_SYMBOLS.VERTICAL + TREE_SYMBOLS.INDENT.slice(1)) ||
+                next.startsWith('|' + TREE_SYMBOLS.INDENT.slice(1))
+            ) {
                 i += TREE_SYMBOLS.VERTICAL.length + TREE_SYMBOLS.INDENT.length - 1;
                 level++;
             } else if (next.startsWith(TREE_SYMBOLS.INDENT)) {
@@ -24,8 +27,15 @@ class LineParser {
     static extractName(rawLine: string): string {
         let line = rawLine;
         // strip indent groups
-        while (line.startsWith(TREE_SYMBOLS.VERTICAL + TREE_SYMBOLS.INDENT.slice(1)) || line.startsWith('|' + TREE_SYMBOLS.INDENT.slice(1)) || line.startsWith(TREE_SYMBOLS.INDENT)) {
-            if (line.startsWith(TREE_SYMBOLS.VERTICAL + TREE_SYMBOLS.INDENT.slice(1)) || line.startsWith('|' + TREE_SYMBOLS.INDENT.slice(1))) {
+        while (
+            line.startsWith(TREE_SYMBOLS.VERTICAL + TREE_SYMBOLS.INDENT.slice(1)) ||
+            line.startsWith('|' + TREE_SYMBOLS.INDENT.slice(1)) ||
+            line.startsWith(TREE_SYMBOLS.INDENT)
+        ) {
+            if (
+                line.startsWith(TREE_SYMBOLS.VERTICAL + TREE_SYMBOLS.INDENT.slice(1)) ||
+                line.startsWith('|' + TREE_SYMBOLS.INDENT.slice(1))
+            ) {
                 line = line.slice(TREE_SYMBOLS.VERTICAL.length + TREE_SYMBOLS.INDENT.length - 1);
             } else {
                 line = line.slice(TREE_SYMBOLS.INDENT.length);
@@ -36,7 +46,8 @@ class LineParser {
             line = line.slice(TREE_SYMBOLS.BRANCH.length);
         } else if (line.startsWith(TREE_SYMBOLS.LAST)) {
             line = line.slice(TREE_SYMBOLS.LAST.length);
-        } else if (line.startsWith('|-- ')) { // ASCII compatibility
+        } else if (line.startsWith('|-- ')) {
+            // ASCII compatibility
             line = line.slice(4);
         } else if (line.startsWith('`-- ')) {
             line = line.slice(4);
@@ -56,12 +67,14 @@ export class TreeParser {
 
         const level = LineParser.countIndentation(line);
         const name = LineParser.extractName(line);
-        if (!name) { return null; }
+        if (!name) {
+            return null;
+        }
 
         return {
             name: name.endsWith('/') ? name.slice(0, -1) : name,
             level,
-            isDirectory: name.endsWith('/')
+            isDirectory: name.endsWith('/'),
         };
     }
 }
